==> writing/models/01_biomass_models.Rmd <==
---
title: "Models 1 - Herbivore biomass gradients"
output:
  html_document:
    df_print: paged
    fig_width: 7
    fig_height: 6
    fig_caption: true
    css: style.css
    theme: sandstone
  html_notebook: default
  pdf_document: default
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.pos = 'H')  ## Hold figure positions
```

#### Report contains model outputs and diagnostics for grazing biomass models.

-----

&nbsp;

```{r, echo=FALSE, include=FALSE, messages=FALSE}

# package loads
library(ggplot2); library(visreg); library(lme4); library(dplyr); library(tidyr); library(scales); library(sjPlot); library(gridExtra); 
library(funk)
theme_set(theme_bw())

## labelling function
add_label <- function(xfrac, yfrac, label, pos = 4, ...){
  u <- par("usr")
  x <- u[1] + xfrac * (u[2] - u[1])
  y <- u[4] - yfrac * (u[4] - u[3])
  text(x, y, label, pos = pos, ...)
}

# data load
load("../../data/wio_herb_benthic_merged.Rdata")


## assign seychelles 2017 with mean complexity values for now - needs fixed
pred$complexity[pred$dataset == 'Seychelles' & pred$date == 2017] <- mean(pred$complexity)


# estimate mean biomass per site per FG
h <- pred %>% 
  ## sum biomass per FG in each transect
        group_by(dataset, date, reef, site, management, transect, 
                 unique.id, depth, FG,
                         hard.coral, macroalgae, rubble, substrate, complexity, fish.biom) %>%
          summarise(biom = sum(biomass.kgha)) %>%
  ## mean FG biomass across transects at each site
          group_by(dataset, date, reef, site, management, unique.id, depth, FG,
                         hard.coral, macroalgae,  rubble, substrate, complexity, fish.biom) %>%
          summarise(biom = mean(biom)) 


h<- spread(h, FG, biom, fill=0)
colnames(h)[14:16]<-c('browser', 'grazer', 'scraper')
h<-as.data.frame(h)


```

#### Predictor dataframe
***

We expect herbivore biomass to be predicted by hard coral cover, macroalgal cover, complexity, and fishing pressure. These 4 covariates are not correlated (Figure 1), so models will not be biased by collinearity issues. We are scaling covariates to a mean of 0 and standard deviation of 1, which is recommended for multimodel selection approaches, and allows for meaningful comparisons of effect sizes when covariates are on different scales (e.g. benthic cover % vs. habitat complexity).

<br><br>

Additional covariates to consider may be disturbance history, oceanic productivity, and wave energy (Heenan et al. 2016, Proceedings), but these are probably not available for all UVC sites.

<br><br>

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.cap='Figure 1: Correlation between unscaled explanatory covariates', fig.height=6, fig.width=6, fig.align='center'}
pairs2(data.frame(h$hard.coral, h$macroalgae, h$rubble, h$substrate, h$complexity, h$management, h$fish.biom), 
       lower.panel=panel.smooth2, upper.panel = panel.cor)
```
<br><br>

```{r, echo=FALSE, include=TRUE, messages=FALSE}

## scale vars to keep covariate means = 0. This is helpful for comparing effect sizes when covariates are on different scales.
h$hard.coral <- scale(h$hard.coral)
h$macroalgae <- scale(h$macroalgae)
h$complexity <- scale(h$complexity)
h$rubble <- scale(h$rubble)
h$substrate <- scale(h$substrate)
h$fish.biom <- scale(h$fish.biom)

## make dummy variables
h$fish.dummy<-ifelse(h$management=='Fished', 1, 0)
h$pristine.dummy<-ifelse(h$management=='Unfished', 1, 0)
# we use 2 dummy variables for 3 levels

```

```{r, echo=FALSE, include=FALSE, messages=FALSE}

## convert biomass to log10
h$grazerlog10<-log10(h$grazer+1)
h$scraperlog10<-log10(h$scraper +1)
h$browserlog10<-log10(h$browser+1)

```

\newpage

---- 

# Grazer models
***

Now fitting models to grazers, with scaled continuous covariates and dummy categorical covariates. We will account for non-independence of islands and sites (i.e. correlated biomass estimates within a region) with random effect structures. 

Fitted model is:

grazerlog10 ~ fixed(hard.coral + macroalgae + complexity + substrate + rubble + fished + pristine) + random(dataset/reef)

where grazerlog10 is grazer biomass (log10 scale), fished indicates if a site is protected or fished, and pristine indicates if a site is pristine (i.e. remote Chagos reefs). Random effect is varying intercepts for each reef, nested in each dataset.

<br><br>

```{r, echo=FALSE, include=FALSE, messages=FALSE}
#### GRAZER BIOMASS MODELS

## drop 0 grazer site
h<-h[h$grazer > 0,]

m.grazer<-lmer(grazerlog10 ~ hard.coral + macroalgae + rubble + substrate + complexity + fish.biom + fish.dummy + pristine.dummy + ## fixed
          (1 | dataset/reef) , ## random, nested = reefs within datasets
                data = h)
#print(summary(m.grazer),  digits = 2, signif.stars = FALSE)

save(m.grazer, file = '../../results/models/biomass_m.grazers.Rdata')

```

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align = 'center', fig.cap='Figure 2: Grazer model diagnostics - residuals and fitted values'}
## model diagnostic plots - grazer biomass
par(mfrow=c(2,2), mar=c(5,4,2,2))
hist(resid(m.grazer), main = 'model residuals', xlab='Residuals', cex.main=0.8)
plot(resid(m.grazer), fitted(m.grazer), xlab='Residuals', ylab= 'Fitted values', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(resid(m.grazer), h$grazerlog10, xlab='Residuals', ylab= 'Observed biomass', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(predict(m.grazer), h$grazerlog10, xlab='Predicted biomass', ylab='Observed biomass', pch=16, col=alpha('black', 0.5), cex.main=0.8)

```


<br><br>

Model residuals are approx. normal (though fail a shapiro test), while model predictions are reasonable. Note clustering of predictions around 1-2 observed biomass. Suggests that commonly-observed biomass values are not well predicted by habitat or fishing. Also note positive residual relationship, indicating that small biomass values are underpredicted and large biomass are overpredicted. 

<br><br>

Next step, examing model estimated effects by visualizing change in biomass across observed gradients in explanatory covariates.

\newpage

### Benthic effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

## NOW PLOT PREDICTIONS

## defining colours for all plots
#install.packages("wesanderson")
library(wesanderson); pal <- wes_palette("Zissou1", 21, type = "continuous")
cols<-c(pal[5], pal[12], pal[18], pal[1], pal[15],pal[21] )


## this dataframe can be used for predictions. We generate 30 values of each continous covariate, ranging from the smallest to largest observed value. This dataset holds fishing effects constant by setting variable to their means (= 0).

## Note that to use the prediction function here, the variable names in pred.master must match the variable names in the fitted dataset (the one called 'h').
cont.pred.master<-data.frame(hard.coral = seq(min(h$hard.coral), max(h$hard.coral), length.out=30),
                             macroalgae = seq(min(h$macroalgae), max(h$macroalgae), length.out=30),
                             complexity = seq(min(h$complexity), max(h$complexity), length.out=30),
                             rubble = seq(min(h$rubble), max(h$rubble), length.out=30),
                             substrate = seq(min(h$substrate), max(h$substrate), length.out=30),
                             fish.biom = 0,
                   fish.dummy = 0, pristine.dummy=0)

## 1. predict grazer biomass for hard coral effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.coral<-predict(object=m.grazer, newdata = nd, re.form=NA)

## 2. predict grazer biomass for macroalgae effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$hard.coral <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.algae<-predict(object=m.grazer, newdata = nd, re.form=NA)

## 3. predict grazer biomass for rubble effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$complexity <- 0; nd$substrate <- 0
p.rubble<-predict(object=m.grazer, newdata = nd, re.form=NA)

## 4. predict grazer biomass for complexity effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$substrate <- 0
p.complex<-predict(object=m.grazer, newdata = nd, re.form=NA)

## 5. predict grazer biomass for substrate effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$complexity <- 0
p.substrate<-predict(object=m.grazer, newdata = nd, re.form=NA)

```

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap = 'Figure 3: Benthic effects on grazers', fig.width=9, fig.height=4}

## create plot for benthic cover predictions. We will add all benthic predictions to the same figure panel using multiple panels
par(mfrow=c(2,3), mar=c(4,2,1,1), oma=c(0,2, 0,0))

## plot coral predictions
plot(cont.pred.master$hard.coral, p.coral, type='l',col = cols[1],lwd=2, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$hard.coral), max(h$hard.coral), length.out=5), 
     labels= round(seq(min(pred$hard.coral), max(pred$hard.coral), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Hard coral')

## plot macroalgal predictions
plot(cont.pred.master$macroalgae, p.algae, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$macroalgae), max(h$macroalgae), length.out=5), 
     labels= round(seq(min(pred$macroalgae), max(pred$macroalgae), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Macroalgae')

## plot rubble predictions
plot(cont.pred.master$rubble, p.rubble, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$rubble), max(h$rubble), length.out=5), 
     labels= round(seq(min(pred$rubble), max(pred$rubble), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Rubble')

## plot complexity predictions
plot(cont.pred.master$complexity, p.complex, type='l',col = cols[3],lwd=2, axes=F, ylab='', xlab='Habitat complexity', ylim=c(1, 2))
axis(1, at = seq(min(h$complexity), max(h$complexity), length.out=5), 
     labels= round(seq(min(pred$complexity), max(pred$complexity), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Complexity')

## plot substrate predictions
plot(cont.pred.master$substrate, p.substrate, type='l',col = cols[1],lwd=2, axes=F, ylab='', xlab='Available substrate', ylim=c(1, 2.5))
axis(1, at = seq(min(h$substrate), max(h$substrate), length.out=5), 
     labels= round(seq(min(pred$substrate), max(pred$substrate), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Substrate')

## label y axis
mtext(2, text= 'Biomass (log10 kg/ha)', outer=TRUE, line=0.5)

```
<br><br>

* Macroalgae has the strongest effect on grazer biomass, which declined from ```r round(10^p.algae[1], 0)``` kg/ha at 0% cover to ```r round(10^p.algae[30], 0)``` kg/ha at ```r round(max(pred$macroalgae), 0)``` % cover.

* Complexity has a moderate effect on grazer biomass, which increased from ```r round(10^p.complex[1], 0)``` kg/ha at 0 complexity to ```r round(10^p.complex[30], 0)``` kg/ha at ```r round(max(pred$complexity), 0)```  complexity.

* Hard coral has the weakest effect on grazer biomass, which increased from ```r round(10^p.coral[1], 0)``` kg/ha at 0% cover to ```r round(10^p.coral[30], 0)``` kg/ha at ```r round(max(pred$hard.coral), 0)``` % cover.

* Rubble has a moderate effect on grazer biomass, which increased from ```r round(10^p.rubble[1], 0)``` kg/ha at 0% cover to ```r round(10^p.rubble[30], 0)``` kg/ha at ```r round(max(pred$rubble), 0)``` % cover.

* Available substrate has a strong effect on grazer biomass, which declined from ```r round(10^p.substrate[1], 0)``` kg/ha at 0 % available substrate to ```r round(10^p.substrate[30], 0)``` kg/ha at ```r round(max(pred$substrate), 0)```  % available substrate.

\newpage

### Fishing effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

# we need a new prediction data frame for the fishing effects, which are categorical.
## Let's use expand.grid to get all combinations of fishing variables, holding benthic covariates to 0
cat.pred.master<-expand.grid(hard.coral = 0,
                             macroalgae = 0,
                             rubble = 0, substrate = 0,
                             complexity = 0, fish.biom = 0, fish.dummy = c(0,1), pristine.dummy=c(0,1))

## Wait. this is wrong. We don't have fished AND pristine sites (duh). Let's drop that row.
cat.pred.master<-cat.pred.master[-4,]

## get fish biom range
fish.master<-data.frame(hard.coral = 0, macroalgae = 0, rubble = 0, complexity = 0, substrate = 0, fish.dummy = 0, pristine.dummy = 0, fish.biom = seq(min(h$fish.biom), max(h$fish.biom), length.out = 30))

## 1. predict grazer biomass for fishing effects
nd<-cat.pred.master
nd$p.fish<-predict(object=m.grazer, newdata = nd, re.form=NA)

## It's easier to add these variables into the predictor variable, so we know which fishing categories they correspond to
nd$categories<-c('Protected', 'Fished', 'Pristine') 

## let's reorder to have a fishing gradient
nd<-nd[c(3,1,2),]

## 2. predict grazer biomass for fishing biomass gradient
fish.master$p.fish<-predict(object=m.grazer, newdata = fish.master, re.form=NA)

```

* Pristine reefs have the highest grazer biomass, at ```r round(10^nd$p.fish[nd$categories=='Pristine'], 0)``` kg/ha.

* Protected and fished reefs have similar biomass levels, at ```r round(10^nd$p.fish[nd$categories=='Protected'], 0)``` kg/ha on protected reefs and ```r round(10^nd$p.fish[nd$categories=='Fished'], 0)``` kg/ha on fished reefs.

* As expected, grazer biomass increases with fishable biomass.

<br><br>

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap='Figure 4: fishing effects on grazers', fig.width=9, fig.height=4}
## 2. create plot for fishing predictions. 
par(mfrow=c(1,2), mar=c(4,4,1,1), oma=c(0,0,0,0))
plot(c(1,2,3), nd$p.fish, cex=2, col = cols[c(4,5,6)],pch=16, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='', ylim=c(1, 3))
axis(1, at = c(1,2,3), labels= nd$categories)
axis(2)

plot(fish.master$fish.biom, fish.master$p.fish, cex=2, axes=F, ylab='', xlab='Fishable biomass (kg/ha)', ylim=c(1, 2.5), type='l')
axis(1, at = seq(min(h$fish.biom), max(h$fish.biom), length.out=5), 
     labels= comma(round(seq(min(pred$fish.biom), max(pred$fish.biom), length.out=5), 0)))
axis(2)


```

<br><br>

\newpage

----

# Browser models
&nbsp;


```{r, echo=FALSE, include=FALSE, messages=FALSE}
#### BROWSER BIOMASS MODELS

## drop 0 grazer site
h<-h[h$browser > 0,]

m.browser<-lmer(browserlog10 ~ hard.coral + macroalgae + complexity + rubble + substrate + fish.biom + fish.dummy + pristine.dummy + ## fixed
          (1 | dataset/reef) , ## random, nested = reefs within datasets
                data = h)
print(summary(m.browser),  digits = 2, signif.stars = FALSE)
save(m.browser, file = '../../results/models/biomass_m.browsers.Rdata')
```

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align = 'center', fig.cap='Figure 5: Browser model diagnostics - residuals and fitted values'}
## model diagnostic plots - grazer biomass
par(mfrow=c(2,2), mar=c(5,4,2,2))
hist(resid(m.browser), main = 'model residuals', xlab='Residuals', cex.main=0.8)
plot(resid(m.browser), fitted(m.browser), xlab='Residuals', ylab= 'Fitted values', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(resid(m.browser), h$browserlog10, xlab='Residuals', ylab= 'Observed biomass', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(predict(m.browser), h$browserlog10, xlab='Predicted biomass', ylab='Observed biomass', pch=16, col=alpha('black', 0.5), cex.main=0.8)

```

<br><br>

Model residuals are normal, not so sure about model predictions, they are okay - similar to grazer model. Still clustering in predictions. Positive residual relationship. Probably should do something with chagos data.


<br><br>

\newpage

### Benthic effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

## NOW PLOT PREDICTIONS

## Note that to use the prediction function here, the variable names in pred.master must match the variable names in the fitted dataset (the one called 'h').
cont.pred.master<-data.frame(hard.coral = seq(min(h$hard.coral), max(h$hard.coral), length.out=30),
                             macroalgae = seq(min(h$macroalgae), max(h$macroalgae), length.out=30),
                             complexity = seq(min(h$complexity), max(h$complexity), length.out=30),
                             rubble = seq(min(h$rubble), max(h$rubble), length.out=30),
                             substrate = seq(min(h$substrate), max(h$substrate), length.out=30),
                             fish.biom = 0,
                   fish.dummy = 0, pristine.dummy=0)

## 1. predict browser biomass for hard coral effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.coral<-predict(object=m.browser, newdata = nd, re.form=NA)

## 2. predict browser biomass for macroalgae effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$hard.coral <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.algae<-predict(object=m.browser, newdata = nd, re.form=NA)

## 3. predict browser biomass for rubble effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$complexity <- 0; nd$substrate <- 0
p.rubble<-predict(object=m.browser, newdata = nd, re.form=NA)

## 4. predict browser biomass for complexity effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$substrate <- 0
p.complex<-predict(object=m.browser, newdata = nd, re.form=NA)

## 5. predict browser biomass for substrate effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$complexity <- 0
p.substrate<-predict(object=m.browser, newdata = nd, re.form=NA)

```

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap = 'Figure 6: Benthic effects on browsers', fig.width=9, fig.height=6}

## create plot for benthic cover predictions. We will add all benthic predictions to the same figure panel using multiple panels
par(mfrow=c(2,2), mar=c(4,2,1,1), oma=c(0,2, 0,0))


## create plot for benthic cover predictions. We will add all benthic predictions to the same figure panel using multiple panels
par(mfrow=c(2,3), mar=c(4,2,1,1), oma=c(0,2, 0,0))

## plot coral predictions
plot(cont.pred.master$hard.coral, p.coral, type='l',col = cols[1],lwd=2, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$hard.coral), max(h$hard.coral), length.out=5), 
     labels= round(seq(min(pred$hard.coral), max(pred$hard.coral), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Hard coral')

## plot macroalgal predictions
plot(cont.pred.master$macroalgae, p.algae, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$macroalgae), max(h$macroalgae), length.out=5), 
     labels= round(seq(min(pred$macroalgae), max(pred$macroalgae), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Macroalgae')

## plot rubble predictions
plot(cont.pred.master$rubble, p.rubble, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(1, 2))
axis(1, at = seq(min(h$rubble), max(h$rubble), length.out=5), 
     labels= round(seq(min(pred$rubble), max(pred$rubble), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Rubble')

## plot complexity predictions
plot(cont.pred.master$complexity, p.complex, type='l',col = cols[3],lwd=2, axes=F, ylab='', xlab='Habitat complexity', ylim=c(1, 2))
axis(1, at = seq(min(h$complexity), max(h$complexity), length.out=5), 
     labels= round(seq(min(pred$complexity), max(pred$complexity), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Complexity')

## plot substrate predictions
plot(cont.pred.master$substrate, p.substrate, type='l',col = cols[1],lwd=2, axes=F, ylab='', xlab='Available substrate', ylim=c(1, 2.5))
axis(1, at = seq(min(h$substrate), max(h$substrate), length.out=5), 
     labels= round(seq(min(pred$substrate), max(pred$substrate), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Substrate')

## label y axis
mtext(2, text= 'Biomass (log10 kg/ha)', outer=TRUE, line=0.5)

```
<br><br>

* Hard coral has the strongest effect on browser biomass, which declined from ```r round(10^p.coral[1], 0)```(52) kg/ha at 0% cover to ```r round(10^p.coral[30], 0)``` (27)kg/ha at ```r round(max(pred$hard.coral), 0)```(92) % cover.

* Macroalgae has a moderate effect on browser biomass, which increased from ```r round(10^p.algae[1], 0)``` kg/ha at 0 complexity to ```r round(10^p.algae[30], 0)``` kg/ha at ```r round(max(pred$macroalgae), 0)```  % cover.

* Rubble has a strong effect on browser biomass, which declined from ```r round(10^p.rubble[1], 0)``` kg/ha at 0% cover to ```r round(10^p.rubble[30], 0)``` kg/ha at ```r round(max(pred$rubble), 0)``` % cover.

* Complexity has the strongest effect on browser biomass, which increased from ```r round(10^p.complex[1], 0)``` kg/ha at 0 complexity to ```r round(10^p.complex[30], 0)``` kg/ha at ```r round(max(pred$complexity), 0)```  complexity.

* Available substrate has a moderate effect on browser biomass, which declined from ```r round(10^p.substrate[1], 0)``` kg/ha at 0 % available substrate to ```r round(10^p.substrate[30], 0)``` kg/ha at ```r round(max(pred$substrate), 0)```  % available substrate.

\newpage

### Fishing effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

# we need a new prediction data frame for the fishing effects, which are categorical.
## Let's use expand.grid to get all combinations of fishing variables, holding benthic covariates to 0
cat.pred.master<-expand.grid(hard.coral = 0,
                             macroalgae = 0,
                             rubble = 0, substrate = 0,
                             complexity = 0, fish.biom = 0, fish.dummy = c(0,1), pristine.dummy=c(0,1))

## Wait. this is wrong. We don't have fished AND pristine sites (duh). Let's drop that row.
cat.pred.master<-cat.pred.master[-4,]

## get fish biom range
fish.master<-data.frame(hard.coral = 0, macroalgae = 0, rubble = 0, complexity = 0, substrate = 0, fish.dummy = 0, pristine.dummy = 0, fish.biom = seq(min(h$fish.biom), max(h$fish.biom), length.out = 30))

## 1. predict browser biomass for fishing effects
nd<-cat.pred.master
nd$p.fish<-predict(object=m.browser, newdata = nd, re.form=NA)

## It's easier to add these variables into the predictor variable, so we know which fishing categories they correspond to
nd$categories<-c('Protected', 'Fished', 'Pristine') 

## let's reorder to have a fishing gradient
nd<-nd[c(3,1,2),]

## 2. predict browser biomass for fishing biomass gradient
fish.master$p.fish<-predict(object=m.browser, newdata = fish.master, re.form=NA)

```

* Pristine reefs have the highest browser biomass, at ```r round(10^nd$p.fish[nd$categories=='Pristine'], 0)``` kg/ha.

* Protected and fished reefs have similar biomass levels, at ```r round(10^nd$p.fish[nd$categories=='Protected'], 0)``` kg/ha on protected reefs and ```r round(10^nd$p.fish[nd$categories=='Fished'], 0)``` kg/ha on fished reefs.

* As expected, browser biomass increases with fishable biomass

<br><br>

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap='Figure 7: fishing effects on browsers', fig.width=9, fig.height=4}

## 2. create plot for fishing predictions. 
par(mfrow=c(1,2), mar=c(4,4,1,1), oma=c(0,0,0,0))
plot(c(1,2,3), nd$p.fish, cex=2, col = cols[c(4,5,6)],pch=16, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='', ylim=c(1, 3))
axis(1, at = c(1,2,3), labels= nd$categories)
axis(2)

plot(fish.master$fish.biom, fish.master$p.fish, cex=2, axes=F, ylab='', xlab='Fishable biomass (kg/ha)', ylim=c(1, 2.5), type='l')
axis(1, at = seq(min(h$fish.biom), max(h$fish.biom), length.out=5), 
     labels= comma(round(seq(min(pred$fish.biom), max(pred$fish.biom), length.out=5), 0)))
axis(2)


```

<br><br>

\newpage

# Scraper models
***

<br><br>

```{r, echo=FALSE, include=FALSE, messages=FALSE}
#### SCRAPER BIOMASS MODELS

## drop 0 scraper site
h<-h[h$scraper > 0,]

m.scraper<-lmer(scraperlog10 ~ hard.coral + macroalgae + complexity + rubble + substrate + fish.biom + fish.dummy + pristine.dummy + ## fixed
          (1 | dataset/reef) , ## random, nested = reefs within datasets
                data = h)
print(summary(m.scraper),  digits = 2, signif.stars = FALSE)

save(m.scraper, file = '../../results/models/biomass_m.scrapers.Rdata')
```

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align = 'center', fig.cap='Figure 8: Scraper model diagnostics - residuals and fitted values'}
## model diagnostic plots - scraper biomass
par(mfrow=c(2,2), mar=c(5,4,2,2))
hist(resid(m.scraper), main = 'model residuals', xlab='Residuals', cex.main=0.8)
plot(resid(m.scraper), fitted(m.scraper), xlab='Residuals', ylab= 'Fitted values', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(resid(m.scraper), h$scraperlog10, xlab='Residuals', ylab= 'Observed biomass', pch=16,col=alpha('black', 0.5), cex.main=0.8)
plot(predict(m.scraper), h$scraperlog10, xlab='Predicted biomass', ylab='Observed biomass', pch=16, col=alpha('black', 0.5), cex.main=0.8)

```

<br><br>

Model residuals are normally distributed, slight clustering in fitted vs. residuals plot does not appear alarming. Note clustering of predictions around 2-3 observed biomass, suggesting that commonly-observed biomass values are not well predicted by habitat or fishing. The positive residual relationship, underprediction of small biomass values and overprediction of large biomass values, echoed by the positive between predicted and observed values.

\newpage

### Benthic effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

## Note that to use the prediction function here, the variable names in pred.master must match the variable names in the fitted dataset (the one called 'h').
cont.pred.master<-data.frame(hard.coral = seq(min(h$hard.coral), max(h$hard.coral), length.out=30),
                             macroalgae = seq(min(h$macroalgae), max(h$macroalgae), length.out=30),
                             complexity = seq(min(h$complexity), max(h$complexity), length.out=30),
                             rubble = seq(min(h$rubble), max(h$rubble), length.out=30),
                             substrate = seq(min(h$substrate), max(h$substrate), length.out=30),
                             fish.biom = 0,
                   fish.dummy = 0, pristine.dummy=0)

## 1. predict scraper biomass for hard coral effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.coral<-predict(object=m.scraper, newdata = nd, re.form=NA)

## 2. predict scraper biomass for macroalgae effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$hard.coral <- 0; nd$complexity <-0; nd$rubble <- 0; nd$substrate <- 0
p.algae<-predict(object=m.scraper, newdata = nd, re.form=NA)

## 3. predict scraper biomass for rubble effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$complexity <- 0; nd$substrate <- 0
p.rubble<-predict(object=m.scraper, newdata = nd, re.form=NA)

## 4. predict scraper biomass for complexity effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$substrate <- 0
p.complex<-predict(object=m.scraper, newdata = nd, re.form=NA)

## 5. predict scraper biomass for substrate effect, holding every other covariate constant
nd<-cont.pred.master
## set non-focal benthic covariates to 0
nd$macroalgae <- 0; nd$hard.coral <-0; nd$rubble <- 0; nd$complexity <- 0
p.substrate<-predict(object=m.scraper, newdata = nd, re.form=NA)


save(cont.pred.master, file = '../../results/models/biomass_m.predictions.Rdata')
```


```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap = 'Figure 9: Benthic effects on scrapers', fig.width=9, fig.height=6}

## create plot for benthic cover predictions. We will add all benthic predictions to the same figure panel using multiple panels
par(mfrow=c(2,2), mar=c(4,2,1,1), oma=c(0,2, 0,0))


## create plot for benthic cover predictions. We will add all benthic predictions to the same figure panel using multiple panels
par(mfrow=c(2,3), mar=c(4,2,1,1), oma=c(0,2, 0,0))

## plot coral predictions
plot(cont.pred.master$hard.coral, p.coral, type='l',col = cols[1],lwd=2, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='Benthic cover (%)', ylim=c(2, 3))
axis(1, at = seq(min(h$hard.coral), max(h$hard.coral), length.out=5), 
     labels= round(seq(min(pred$hard.coral), max(pred$hard.coral), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Hard coral')

## plot macroalgal predictions
plot(cont.pred.master$macroalgae, p.algae, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(2, 3))
axis(1, at = seq(min(h$macroalgae), max(h$macroalgae), length.out=5), 
     labels= round(seq(min(pred$macroalgae), max(pred$macroalgae), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Macroalgae')

## plot rubble predictions
plot(cont.pred.master$rubble, p.rubble, type='l',col = cols[2],lwd=2, axes=F, ylab='', xlab='Benthic cover (%)', ylim=c(2, 3))
axis(1, at = seq(min(h$rubble), max(h$rubble), length.out=5), 
     labels= round(seq(min(pred$rubble), max(pred$rubble), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Rubble')

## plot complexity predictions
plot(cont.pred.master$complexity, p.complex, type='l',col = cols[3],lwd=2, axes=F, ylab='', xlab='Habitat complexity', ylim=c(2, 3))
axis(1, at = seq(min(h$complexity), max(h$complexity), length.out=5), 
     labels= round(seq(min(pred$complexity), max(pred$complexity), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Complexity')

## plot substrate predictions
plot(cont.pred.master$substrate, p.substrate, type='l',col = cols[1],lwd=2, axes=F, ylab='', xlab='Available substrate', ylim=c(2, 3))
axis(1, at = seq(min(h$substrate), max(h$substrate), length.out=5), 
     labels= round(seq(min(pred$substrate), max(pred$substrate), length.out=5), 0))
axis(2)
add_label(0.01, 0.1, font = 2, label ='Substrate')

## label y axis
mtext(2, text= 'Biomass (log10 kg/ha)', outer=TRUE, line=0.5)

```
<br><br>

* Hard coral has a moderate effect on scraper biomass, which increased from ```r round(10^p.coral[1], 0)``` kg/ha at 0% cover to ```r round(10^p.coral[30], 0)``` kg/ha at ```r round(max(pred$hard.coral), 0)``` % cover.

* Macroalgae has the weakest effect on scraper biomass, which declined from ```r round(10^p.algae[1], 0)``` kg/ha at 0% cover to ```r round(10^p.algae[30], 0)``` kg/ha at ```r round(max(pred$macroalgae), 0)``` % cover.

* Rubble has a moderate effect on scraper biomass, which increased from ```r round(10^p.rubble[1], 0)``` kg/ha at 0 complexity to ```r round(10^p.rubble[30], 0)``` kg/ha at ```r round(max(pred$rubble), 0)```  % cover.

* Complexity has the strongest effect on scraper biomass, which increased from ```r round(10^p.complex[1], 0)``` kg/ha at 0 complexity to ```r round(10^p.complex[30], 0)``` kg/ha at ```r round(max(pred$complexity), 0)```  complexity.

* Available substrate has a strong effect on scraper biomass, which increased from ```r round(10^p.substrate[1], 0)``` kg/ha at 0 % available substrate to ```r round(10^p.substrate[30], 0)``` kg/ha at ```r round(max(pred$substrate), 0)```  % available substrate.

\newpage

### Fishing effects
***

```{r, echo=FALSE, include=FALSE, messages=FALSE}

# we need a new prediction data frame for the fishing effects, which are categorical.
## Let's use expand.grid to get all combinations of fishing variables, holding benthic covariates to 0
cat.pred.master<-expand.grid(hard.coral = 0,
                             macroalgae = 0,
                             rubble = 0, substrate = 0,
                             complexity = 0, fish.biom = 0, fish.dummy = c(0,1), pristine.dummy=c(0,1))

## Wait. this is wrong. We don't have fished AND pristine sites (duh). Let's drop that row.
cat.pred.master<-cat.pred.master[-4,]

## get fish biom range
fish.master<-data.frame(hard.coral = 0, macroalgae = 0, rubble = 0, complexity = 0, substrate = 0, fish.dummy = 0, pristine.dummy = 0, fish.biom = seq(min(h$fish.biom), max(h$fish.biom), length.out = 30))

## 1. predict scraper biomass for fishing effects
nd<-cat.pred.master
nd$p.fish<-predict(object=m.scraper, newdata = nd, re.form=NA)

## It's easier to add these variables into the predictor variable, so we know which fishing categories they correspond to
nd$categories<-c('Protected', 'Fished', 'Pristine') 

## let's reorder to have a fishing gradient
nd<-nd[c(3,1,2),]

## 2. predict scraper biomass for fishing biomass gradient
fish.master$p.fish<-predict(object=m.scraper, newdata = fish.master, re.form=NA)

save(fish.master, file = '../../results/models/biomass_fishing.effect.predictions.Rdata')
```

* Pristine reefs have the highest scraper biomass, at ```r round(10^nd$p.fish[nd$categories=='Pristine'], 0)``` kg/ha.

* Protected reefs have a scraper biomass of ```r round(10^nd$p.fish[nd$categories=='Protected'], 0)``` kg/ha while fished reefs show a scraper biomass of ```r round(10^nd$p.fish[nd$categories=='Fished'], 0)``` kg/ha.

* Fishable biomass for scrapers ranges from ```r round(range(fish.master$p.fish[fish.master$model == 'grazers']), digits =0) ``` 

<br><br>

```{r, echo=FALSE, include=TRUE, messages=FALSE, fig.align='center', fig.cap='Figure 10: fishing effects on scrapers', fig.width=9, fig.height=4}

## 2. create plot for fishing predictions. 
par(mfrow=c(1,2), mar=c(4,4,1,1), oma=c(0,0,0,0))
plot(c(1,2,3), nd$p.fish, cex=2, col = cols[c(4,5,6)],pch=16, axes=F, ylab='Log10 Biomass (kg/ha)', xlab='', ylim=c(2, 3))
axis(1, at = c(1,2,3), labels= nd$categories)
axis(2)

plot(fish.master$fish.biom, fish.master$p.fish, cex=2, axes=F, ylab='', xlab='Fishable biomass (kg/ha)', ylim=c(2,3), type='l')
axis(1, at = seq(min(h$fish.biom), max(h$fish.biom), length.out=5), 
     labels= comma(round(seq(min(pred$fish.biom), max(pred$fish.biom), length.out=5), 0)))
axis(2)


```

<br><br>

\newpage


# Effect sizes across feeding groups

***

Because covariates were scaled and centered, we can directly compare effect sizes among models and among covariates with different scales (e.g. fishable biomass and percent cover). Pristine reefs - Chagos - are the strongest predictor of biomass, with higher biomass of all 3 FGs. These estimates were highly uncertain, indicative of high site-level variability in biomass. Fishable biomass was consistently strong positive influence on FG biomass, as was rubble, which outweighed hard coral cover as a predictor for all FGs.

Macroalgae had positive effects on browser biomass, and negative effects on grazer and scraper biomass, while complexity was negatively associated with all 3 FGs. On inhabited reefs, fishing had significant negative effects on biomass, though the effect of protection was weaker than that of reef isolation (i.e. remote/pristine Chagos).

```{r, include =T, echo = F, messages=F, fig.caption = 'Figure 11: relative effect of fishing and benthic covariates', fig.width = 12, fig.height = 4, fig.align = 'center'}

library(sjlabelled)

## create df for labelling plots with better names
labs<-data.frame(lab=c('Hard coral', 'Macroalgae', 'Rubble', 'Available substrate', 'Complexity', 'Fishable biomass', 'Fished',   'Pristine'),
                 model=c('hard.coral', 'macroalgae', 'rubble', 'substrate','complexity', 'fish.biom', 'fish.dummy', 'pristine.dummy'))

## extract correct order of names according to increasing parameter estimate size
od<-sort(fixef(m.grazer), decreasing=T)
od<-od[-which(names(od) == '(Intercept)')]

# plot fixed param estimates, annotated with values
g1<-plot_model(m.grazer, arg = 'est', pred.type = 'fe', axis.labels = rev(labs$lab[match(names(od), labs$model)]),
           sort.est = T, show.values =T,value.offset = .3) + ggtitle('Grazers')

## repeat for browsers
od<-sort(fixef(m.browser), decreasing=T)
od<-od[-which(names(od) == '(Intercept)')]

g2<-plot_model(m.browser, arg = 'est', pred.type = 'fe', axis.labels = rev(labs$lab[match(names(od), labs$model)]),
           sort.est = T, show.values =T,value.offset = .3) + ggtitle('Browsers')

## repeat for scrapers
od<-sort(fixef(m.scraper), decreasing=T)
od<-od[-which(names(od) == '(Intercept)')]

g3<-plot_model(m.scraper, arg = 'est', pred.type = 'fe', axis.labels = rev(labs$lab[match(names(od), labs$model)]),
           sort.est = T, show.values =T,value.offset = .3) + ggtitle('Scrapers')

grid.arrange(g1, g2, g3, nrow = 1, ncol=3)
```

#### Figure 2

```{r, include =T, echo = F, messages=F, fig.caption = 'Figure 12. Biomass model results: a. Relative effect sizes of each FG. b. macroalgae. c.available substrate. d. fishable biomass.', fig.width = 12, fig.height = 4, fig.align = 'center'}


#panel 1: relative effect sizes of each FG



```

#### Results paragraph 1

Our fitted generalised linear mixed model showed differentiating effects for each functional group.

Increasing cover of macroalgae (Fig. 2a) negatively affected biomasses of both herbivore grazers and scrapers, while it slightly elevated that of browsers. Grazer biomass declined from 59 kg/ha at 0% cover to 26 kg/ha at 51 % cover and scraper biomass declined from 265 kg/ha at 0% macroalgae cover to 216 kg/ha at 51 % cover. Browser biomass increased from 47 kg/ha at 0 % macroalgae cover to 66 kg/ha at 51 % cover.

Increasing cover of available substrate (Fig.2b) positively affected herbivore grazers and scrapers, while it negatively affected brower biomass. Grazer biomass increased from 41 kg/ha at 0 % available substrate to 120 kg/ha at 88 % available substrate and scraper biomass increased from 205 kg/ha at 0 % available substrate to 469 kg/ha at 88 % available substrate. Browser biomass declined from 57 kg/ha at 0 % available substrate to 33 kg/ha at 88 % available substrate.

As biomasses of grazers, scrapers, and browsers increased, so did the fishable biomass (Fig. 2c) of each group. This effect was strongly pronounced for browsers, while that for grazers and scrapers was comparable. Grazer biomass increased from 52 kg/ha to 120 kg/ha and scraper biomass increased from 222 kg/ha to 1582 kg/ha across the fishable biomass gradient. Browser biomass increased from 44 kg/ha to 134 kg/ha across the fishable biomass gradient.

Effect sizes (Fig. 2d) expectedly also varied between the three functional groups. The order in which predictors are mentioned is according to the effect size from strongest to weakest. Grazer biomass was positively affected by pristine reefs, available substrate, fishable biomass, structural complexity, and rubble cover. Macroalgae cover, fished reefs, and hard coral cover negatively affected grazer biomass. Scraper biomass was positively affected by pristine reefs, fishable biomass, available substrate, structural complexity, and rubble cover. Fished reefs, macroalgae cover, and hard coral cover negatively affected scraper biomass. Browser biomass was positively affected by pristine reefs, structural complexity, and rubble cover. Fished reefs, Macroalgae cover, and hard coral cover negatively affected browser biomass.

----
