==> writing/models/02_bite-rate-models.Rmd <==
---
title: "Models 2 - Bite rates and grazing functions"
output:
  html_document:
    df_print: paged
    fig_width: 7
    fig_height: 6
    fig_caption: true
    css: style.css
    theme: sandstone
  html_notebook: default
  pdf_document: default
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.pos = 'H')  ## Hold figure positions
eval.mods=FALSE
```

#### Report contains models for converting grazing biomass to functions

-----

<br><br>

```{r, echo=FALSE, include=F, messages=FALSE}

# package loads
library(ggplot2); library(funk); library(gridExtra); library(dplyr); library(tidyr); library(rethinking); library(bayesplot)
theme_set(theme_sleek())

## use cores
options(mc.cores = parallel::detectCores())

# data load
bite<-read.csv("bite_rate_clean.csv")

```

Bite data were collected by Andy Hoey in the Red Sea, Indonesia, and GBR. Alexia Graba-Landry collected additional acanthurid and siganid/naso observations in the GBR. Data are underwater feeding observations that record the number of bites by herbivorous fish. Across the dataset, ```r dim(bite)[1]``` fishes were observed, including ```r uniques(bite$sp)``` species from ```r uniques(bite$Genus)``` families. Observation times ranged from ```r min(bite$total.sec)/60``` minutes to ```r max(bite$total.sec)/60``` minutes, and averaged ```r round(mean(bite$total.sec)/60, 2)``` minutes.

<br><br>

***

```{r, echo=FALSE, include = TRUE, messages = FALSE, fig.cap='Figure 1: observed bite rates by functional group and region',  fig.align = 'center', fig.height = 5, fig.width = 8}

ggplot(bite, aes(TL, Bite.rate, col = dataset)) + geom_point(alpha=0.5) + facet_wrap(~FG) +
      labs(x = 'Total length (cm)', y = 'Bites per minute')
```
<br><br>

***

Size relationship is apparent for scrapers only, with largest individuals having the lowest bite rates (Figure 1). Grazer bite rates are highly variable among species, while browser rates are lower and less frequently observed. GBR datasets dominate the observations, with Indonesia and Red Sea only contributing to scrapers. 

<br><br>
```{r, echo=FALSE, include = T, messages = FALSE, fig.cap='Figure 2: observed bite rates by functional group and species',  fig.align = 'center', fig.height = 7, fig.width = 12}

ggplot(bite, aes(TL, Bite.rate, col = dataset)) + geom_point(alpha=0.5, size = 0.8) + facet_wrap(~sp, scales = 'free_y') +
      labs(x = 'Total length (cm)', y = 'Bites per minute') + 
      theme(legend.position = 'top', strip.text.x = element_text(size = 7))
```

***

After matching these data to UVC species, we have information for 39 species. Therefore, 63 species do not have bite observations and will require family or genus level predictions. These species are 13 browsers, 42 grazers, and 8 scrapers.

<br><br>

```{r, echo =F, messages = F, include=T, fig.cap = 'Figure 3: Species with feeding observations vs. UVC species', fig.align='center', fig.width=8, fig.height=4}

sp<-read.csv('../../data/herbivore_species_list.csv')
sp$Bite.Observations<-ifelse(sp$species %in% bite$sp, 'YES', 'NO')
ggplot(sp, aes(FG, fill=Bite.Observations)) + geom_bar(aes(y = ..count..), position='dodge') +
    geom_text(aes(label=..count.., y=..count..), stat = 'count', position=position_dodge(width=0.9), vjust=-0.25) +
    theme(legend.position = 'top') + labs(x = '')

```


<br><br>

Aim is to build a predictive model that estimates the bite rate of each fish in the UVC dataset, given its species and, for scrapers, its size. Model distributions will be set according to the bite rate distributions, which are all left-skewed. When data are continuous and not normally distributed (Figure 4; grazers and scrapers), we'll use the flexible Gamma distribution.

<br><br>

```{r,echo=FALSE, eval = TRUE, messages = FALSE, fig.align='center', fig.cap='Figure 4: histogram of bite rates by functional group', fig.width=7, fig.height = 4}

par(mfrow=c(1,3), mar=c(2,2,2,2))
hist(bite$Bite.rate[bite$FG == 'Herbivore Grazer'], xlab='Bites per minute', main = 'Grazers')
hist(bite$Bite.rate[bite$FG == 'Herbivore Scraper'], xlab='Bites per minute', main = 'Scrapers')
hist(bite$Bite.rate[bite$FG == 'Herbivore Browser'], xlab='Bites per minute', main = 'Browsers')

```


#### Grazer model
***

Grazer function is cropping/consumption of turf algae. We infer this function from the total bite rate of the grazer community, and assume all bites are the same size. Thus, the predictive model estimates bite rates by species and families. We can use a hierarchical Bayesian model to predict new levels (i.e. species), based on their genus.

<br><br>

$bite = Gamma(\mu, \theta)$

$log(\mu) = X + species_i + genus_j$

<br><br>

where X is the intercept, for ```r uniques(bite$sp[bite$FG == 'Herbivore Grazer'])``` species in ```r uniques(bite$Genus[bite$FG == 'Herbivore Grazer'])``` genera

<br><br>


```{r, echo=FALSE, eval = FALSE, messages = FALSE}

grazers<-droplevels(bite[bite$FG == 'Herbivore Grazer',])
colnames(grazers)[colnames(grazers) == 'Bite.rate']<-'biterate'
bite.prior=mean(grazers$biterate) ## 30.85, log = 3.43

graze.m<-map2stan(
        alist(
          biterate ~ dgamma2(mu, scale),
          log(mu) ~ a + X1[sp] + X2[Genus] + X3[dataset], 
          X1[sp] ~ dnorm(0, sigmar),
          X2[Genus] ~ dnorm(0, sigmar2),
          X3[dataset] ~ dnorm(0, sigmar3),
          a ~ dnorm(3.43, 10),
          scale ~ dexp(2),
          c(sigmar, sigmar2, sigmar3) ~ dcauchy(0, 1)
        ),
        data=grazers, iter = 3000, chains =1, cores = 4)

save(grazers, graze.m, file = '../../results/models/bites_grazers.Rdata')
```

```{r, echo=FALSE, eval = TRUE, messages = FALSE, include = FALSE}

## evaluating predictions for each species
load('../../results/models/bites_grazers.Rdata')

## Predictions for species, with genus effect
d.pred <-data.frame(sp = unique(grazers$sp), dataset = 'NA')
d.pred$Genus<-grazers$Genus[match(d.pred$sp, grazers$sp)]
a.dataset.zero = matrix(0, 1000, 3)

link.obs.sp<-link(graze.m, n = 1000, data = as.list(d.pred), replace = list(X3 = a.dataset.zero))
pred.mean<-data.frame(median = apply(link.obs.sp, 2, median))
pred.PI<-apply(link.obs.sp, 2, PI, prob = 0.95)
pred.mean$upper<-pred.PI[2,]
pred.mean$lower<-pred.PI[1,]
pred.mean$class <- d.pred$sp #pred.mean$Genus <- d.pred$Genus
pred.mean$class<- factor(pred.mean$class, level = rev(levels(pred.mean$class)))
pred.mean$preds<-'Species'

## Predictions for genus, removing species effect
d.pred <- data.frame(Genus = unique(grazers$Genus), sp = 'NA', dataset = 'NA')
a.sp.zero = matrix(0, 1000, 9)
a.dataset.zero = matrix(0, 1000, 3)

link.obs.genus<-link(graze.m, n = 1000, data = as.list(d.pred), replace= list(X1= a.sp.zero, X3 = a.dataset.zero))
pred.mean.genus<-data.frame(median = apply(link.obs.genus, 2, median))
pred.PI<-apply(link.obs.genus, 2, PI, prob = 0.95)
pred.mean.genus$upper<-pred.PI[2,]
pred.mean.genus$lower<-pred.PI[1,]
pred.mean.genus$class <- d.pred$Genus
pred.mean.genus$class<- factor(pred.mean.genus$class, level = rev(levels(pred.mean.genus$class)))
pred.mean.genus$preds<-'Genus'

## Predictions for global mean, removing species and genus effect
d.pred <- data.frame(Genus = 'NA', sp = 'NA', dataset = 'NA')
a.sp.zero = matrix(0, 1000, 9)
a.genus.zero = matrix(0, 1000, 5)
a.dataset.zero = matrix(0, 1000, 3)

link.obs.mean<-link(graze.m, n = 1000, data = as.list(d.pred), replace= list(X1= a.sp.zero, X2 = a.genus.zero, X3 = a.dataset.zero))
pred.global<-data.frame(median = apply(link.obs.mean, 2, median))
pred.PI<-apply(link.obs.mean, 2, PI, prob = 0.95)
pred.global$upper<-pred.PI[2,]
pred.global$lower<-pred.PI[1,]
pred.global$class <- 'global.mean'
pred.global$preds<-'global.mean'

p<-rbind(pred.mean, pred.mean.genus, pred.global)
p$col<-grazers$mean[match(p$class, grazers$sp)]
p$col[p$preds == 'Genus']<-grazers$col[match(p$class[p$preds == 'Genus'], grazers$mean)]
p$col[p$preds == 'Genus']<-'black'

write.csv(p, file = '../../results/functions/grazer_bites_predicted.csv')
```

```{r,echo=FALSE,fig.align='center', fig.cap = 'Figure 5: Grazer bite rates by genera and species. Points are median posterior predictions, with 95% certainty intervals.', fig.width= 8, fig.height = 5}

ggplot(p, aes(class, median, col=col)) + geom_pointrange(aes(ymin = lower, ymax = upper)) + coord_flip() + theme(legend.position = 'NULL') + labs( x = '', y = 'Bites per minute', title = 'Grazers') + facet_wrap(~preds, scales='free_y')

```

***

<br><br>

Species-level predictions are very tight, but genera have higher uncertainty (Figure 5). Interesting to see how these compare with browser estimates - from a prediction perspective, is there any benefit in separating these groups? Function is determined by bite volumes and algal specificity, so obviously need to keep functions separate.

<br><br>

#### Scraper model
***

Scraper function is removal of dead coral substrate, endolithic material, and detritus. We infer this function from the total bite rate of the scraper community, and accounting for changes in bite scar volume with size. Thus, the predictive model estimates bite rates by species, genera and size. We can use a hierarchical Bayesian model to predict new levels (i.e. species), based on their genus.

<br><br>

$bite = Gamma(\mu, \theta)$

$log(\mu) = X + B*size + species_i + genus_j$

<br><br>

where X is the intercept and B is the size effect, for ```r uniques(bite$sp[bite$FG == 'Herbivore Scraper'])``` species in ```r uniques(bite$Genus[bite$FG == 'Herbivore Scraper'])``` genera.

<br><br>


```{r, echo=FALSE, eval = eval.mods, messages = FALSE}

scrapers<-droplevels(bite[bite$FG == 'Herbivore Scraper',])
colnames(scrapers)[colnames(scrapers) == 'Bite.rate']<-'biterate'
bite.prior=mean(scrapers$biterate) ## 22.15, log = 3.10

scrape.m<-map2stan(
        alist(
          biterate ~ dgamma2(mu, scale),
          log(mu) ~ a + B*TL + X1[sp] + X2[Genus] + X3[dataset], 
          X1[sp] ~ dnorm(0, sigmar),
          X2[Genus] ~ dnorm(0, sigmar2),
          X3[dataset] ~ dnorm(0, sigmar3),
          a ~ dnorm(3.10, 10),
          B ~ dnorm(0, 5),
          scale ~ dexp(1),
          c(sigmar, sigmar2, sigmar3) ~ dcauchy(0, 1)
        ),
        data=scrapers, warmup = 1500, iter = 5000, chains =1, cores = 4)

save(scrapers, scrape.m, file = '../../results/models/bites_scrapers.Rdata')

## now model for area scraped
area<-read.csv('../../data/bite-rates/scrape_sizes.csv')
## drop zeroes
area <- area[area$bitearea > 0,]
area.prior=mean(area$bitearea) ## 85.67, log = 4.45

scrape.m2<-map2stan(
        alist(
          bitearea ~ dgamma2(mu, scale),
          log(mu) ~ a + B*TL, #+ X1[species] + X2[genus],
          #X1[species] ~ dnorm(0, sigmar),
          #X2[genus] ~ dnorm(0, sigmar2),
          a ~ dnorm(4.45, 5),
          B ~ dnorm(0, 2),
          scale ~ dexp(5)
          #c(sigmar, sigmar2) ~ dcauchy(0, 1)
        ),
        data=area, warmup = 1500, iter = 5000, chains =1, cores = 4)

save(area, scrape.m2, file = '../../results/models/area_scrapers.Rdata')

```

***

For scrapers, there is strong variation in bite rates among species, but less among genera (Figure 6). This suggests that predicting unobserved species may be easier for this FG, though scrapers were well covered for bite rate data. Scarus have the highest growth rates, with some species exceeding 40 bites per minute (S. spinus, globiceps, and viridifucatus). Correcting these rates to volumes grazed is important, so we shouldn't interpret these patterns yet.

<br><br>

```{r, echo=FALSE, eval = TRUE, messages = FALSE, include =FALSE}

## evaluating predictions for each species
load('../../results/models/bites_scrapers.Rdata')

## Predictions for species, with genus effect
d.pred <-data.frame(sp = unique(scrapers$sp), TL = mean(scrapers$TL), dataset = 'NA')
d.pred$Genus<-scrapers$Genus[match(d.pred$sp, scrapers$sp)]
a.dataset.zero = matrix(0, 1000, 3)

link.obs.sp<-link(scrape.m, n = 1000, data = as.list(d.pred), replace=list(X3 = a.dataset.zero))
pred.mean<-data.frame(median = apply(link.obs.sp, 2, median))
pred.PI<-apply(link.obs.sp, 2, PI, prob = 0.95)
pred.mean$upper<-pred.PI[2,]
pred.mean$lower<-pred.PI[1,]
pred.mean$class <- d.pred$sp #pred.mean$Genus <- d.pred$Genus
pred.mean$class<- factor(pred.mean$class, level = rev(levels(pred.mean$class)))
pred.mean$preds<-'Species'

## Predictions for genus, removing species effect
d.pred <- data.frame(Genus = unique(scrapers$Genus), sp = 'NA', TL = mean(scrapers$TL), dataset = 'NA')
a.sp.zero = matrix(0, 1000, 27)
link.obs.genus<-link(scrape.m, n = 1000, data = as.list(d.pred), replace= list(X1= a.sp.zero, X3 = a.dataset.zero))
pred.mean.genus<-data.frame(median = apply(link.obs.genus, 2, median))
pred.PI<-apply(link.obs.genus, 2, PI, prob = 0.95)
pred.mean.genus$upper<-pred.PI[2,]
pred.mean.genus$lower<-pred.PI[1,]
pred.mean.genus$class <- d.pred$Genus
pred.mean.genus$class<- factor(pred.mean.genus$class, level = rev(levels(pred.mean.genus$class)))
pred.mean.genus$preds<-'Genus'

## Predictions for global mean, removing species and genus effect
d.pred <- data.frame(Genus = 'NA', sp = 'NA', dataset = 'NA',TL = mean(scrapers$TL))
a.sp.zero = matrix(0, 1000, 27)
a.genus.zero = matrix(0, 1000, 4)
a.dataset.zero = matrix(0, 1000, 3)

link.obs.mean<-link(scrape.m, n = 1000, data = as.list(d.pred), replace= list(X1= a.sp.zero, X2 = a.genus.zero, X3 = a.dataset.zero))
pred.global<-data.frame(median = apply(link.obs.mean, 2, median))
pred.PI<-apply(link.obs.mean, 2, PI, prob = 0.95)
pred.global$upper<-pred.PI[2,]
pred.global$lower<-pred.PI[1,]
pred.global$class <- 'global.mean'
pred.global$preds<-'global.mean'

p<-rbind(pred.mean, pred.mean.genus, pred.global)
p$col<-scrapers$Genus[match(p$class, scrapers$sp)]
p$col[p$preds == 'Genus']<-scrapers$Genus[match(p$class[p$preds == 'Genus'], scrapers$Genus)]

write.csv(p, file = '../../results/functions/scraper_bites_predicted.csv')

```

```{r, echo=F, fig.align='center', fig.cap = 'Figure 6: Scraper bite rates by genera and species. Points are median posterior predictions, with 95% certainty intervals.', fig.width= 8, fig.height = 5}

ggplot(p, aes(class, median, col=col)) + geom_pointrange(aes(ymin = lower, ymax = upper)) + coord_flip() + theme(legend.position = 'NULL') + labs( x = '', y = 'Bites per minute', title = 'Scrapers') + facet_wrap(~preds, scales='free_y')

```
***

Size effects were also relative strong, with bite rates decreasing by 50% from the smallest to largest individuals (Figure 7). High uncertainty in this relationship suggests there are important species-specific size differences here, but we have not accounted for these.

<br><br>

```{r, echo=FALSE, eval = TRUE, messages = FALSE, include=F}
## size effects
d.pred <- data.frame(Genus = 'NA', sp = 'NA', TL = seq(min(scrapers$TL), max(scrapers$TL), length.out =30), dataset = 'NA')
a.sp.zero = matrix(0, 1000, 28); a.genus.zero = matrix(0, 1000, 4); a.dataset.zero = matrix(0, 1000, 2)
link.obs.size<-link(scrape.m, n = 1000, data = as.list(d.pred), 
                      replace= list(X1= a.sp.zero, X2 = a.genus.zero, X3 = a.dataset.zero))
pred.mean.size<-data.frame(median = apply(link.obs.size, 2, median))
pred.PI<-apply(link.obs.size, 2, PI, prob = 0.95)
pred.mean.size$upper<-pred.PI[2,]
pred.mean.size$lower<-pred.PI[1,]
pred.mean.size$TL<-d.pred$TL

```

```{r,echo=F, fig.align='center', fig.cap = 'Figure 7: Scraper bite rates by size. Lines are median posterior predictions with 95% certainty intervals.', fig.width= 8, fig.height = 4}

ggplot(pred.mean.size, aes(TL, median)) + geom_line() + 
      geom_line(data = pred.mean.size, aes(TL, upper), linetype='dotted') +
      geom_line(data = pred.mean.size, aes(TL, lower), linetype='dotted') + 
      labs(x = 'Total length (cm)', y = 'Bites per minute')

```

***

For bite area, predictions made from separate dataset of bite width and length for scrapers on the GBR. Although data were collected for 17 species of 3 genera, the majority of observations are for Ctetoscarus bicolor, Scarus frenatus, Scarus microrhinos, and Chlorurus sordidus. So, only fitting a size model, ignoring species effects.

```{r, echo=FALSE, eval = TRUE, messages = FALSE, include =FALSE}

## evaluating predictions for each species
load('../../results/models/area_scrapers.Rdata')

## size effects
d.pred <- data.frame(TL = seq(min(area$TL), max(area$TL), length.out =30))
link.obs.size<-link(scrape.m2, n = 1000, data = as.list(d.pred))
pred.mean.size<-data.frame(median = apply(link.obs.size, 2, median))
pred.PI<-apply(link.obs.size, 2, PI, prob = 0.95)
pred.mean.size$upper<-pred.PI[2,]
pred.mean.size$lower<-pred.PI[1,]
pred.mean.size$TL<-d.pred$TL

write.csv(pred.mean.size, file = '../../results/functions/scraper_area_predicted.csv')

```

```{r,echo=F, fig.align='center', fig.cap = 'Figure 8: Scraper bite area by size. Lines are median posterior predictions with 95% certainty intervals.', fig.width= 8, fig.height = 4}

ggplot(pred.mean.size, aes(TL, median)) + geom_line() + 
      geom_line(data = pred.mean.size, aes(TL, upper), linetype='dotted') +
      geom_line(data = pred.mean.size, aes(TL, lower), linetype='dotted') + 
      labs(x = 'Total length (cm)', y = 'Area scraped (mm2)')

```

***

#### Browser model

Browser function is removal/consumption of macroalgae. We infer this function from the total bite rate of the browser community, and assume all bites are the same size. Thus, the predictive model estimates bite rates by 2 genera. We can use a hierarchical Bayesian model to predict new levels (i.e. species), based on their genus. Note there are only 2 species and 2 genera, so here I am fitting a genera model, and we are missing observations of any browsing siganid species (these are high biomass on macroalgal reefs).

<br><br>

$bite = Normal(\mu, \sigma)$

$\mu = X + genus_j$

<br><br>

where X is the intercept, for ```r uniques(bite$sp[bite$FG == 'Herbivore Browser'])``` species in ```r uniques(bite$Genus[bite$FG == 'Herbivore Browser'])``` genera. Species effects are equivalent to genera effects, so only one taxonomic term is fitted ($genus_j$).

<br><br>


```{r, echo=FALSE, eval = eval.mods, messages = FALSE}

browsers<-droplevels(bite[bite$FG == 'Herbivore Browser',])
colnames(browsers)[colnames(browsers) == 'Bite.rate']<-'biterate'
bite.prior=mean(browsers$biterate) ## 11.98

browse.m<-map2stan(
        alist(
          biterate ~ dnorm(mu, sigma),
          mu ~ a + X2[Genus] + X3[dataset], 
          #X1[sp] ~ dnorm(0, sigmar),
          X2[Genus] ~ dnorm(0, sigmar2),
          X3[dataset] ~ dnorm(0, sigmar3),
          a ~ dnorm(11.98, 10),
          c(sigma, sigmar2, sigmar3) ~ dcauchy(0, 1)
        ),
        data=browsers, iter = 3000, chains =1, cores = 4)

save(browsers, browse.m, file = '../../results/models/bites_browsers.Rdata')
```

```{r, echo=FALSE, eval = TRUE, messages = FALSE, include=F}

## evaluating predictions for each species
load('../../results/models/bites_browsers.Rdata')

## Predictions for genus
d.pred <- data.frame(Genus = unique(browsers$Genus), dataset = 'NA')
link.obs.genus<-link(browse.m, n = 1000, data = as.list(d.pred))
pred.mean.genus<-data.frame(median = apply(link.obs.genus, 2, median))
pred.PI<-apply(link.obs.genus, 2, PI, prob = 0.95)
pred.mean.genus$upper<-pred.PI[2,]
pred.mean.genus$lower<-pred.PI[1,]
pred.mean.genus$class <- d.pred$Genus
pred.mean.genus$class<- factor(pred.mean.genus$class, level = rev(levels(pred.mean.genus$class)))
pred.mean.genus$preds<-'Genus'

## Predictions for global mean, removing species and genus effect
d.pred <- data.frame(Genus = 'NA', dataset = 'NA')
a.genus.zero = matrix(0, 1000, 2)
a.dataset.zero = matrix(0, 1000, 3)

link.obs.mean<-link(browse.m, n = 1000, data = as.list(d.pred), replace= list(X2 = a.genus.zero, X3 = a.dataset.zero))
pred.global<-data.frame(median = apply(link.obs.mean, 2, median))
pred.PI<-apply(link.obs.mean, 2, PI, prob = 0.95)
pred.global$upper<-pred.PI[2,]
pred.global$lower<-pred.PI[1,]
pred.global$class <- 'global.mean'
pred.global$preds<-'global.mean'

p<-rbind(pred.mean.genus, pred.global)
p$col<-browsers$mean[match(p$class, browsers$sp)]
p$col[p$preds == 'Genus']<-browsers$col[match(p$class[p$preds == 'Genus'], browsers$mean)]
p$col[p$preds == 'Genus']<-'black'

write.csv(p, file = '../../results/functions/browser_bites_predicted.csv')

```

Browser bite rates are lower than grazers (Fig. 8), but it's difficult to infer much with only 2 species. We can't, for example, extrapolate these to siganid species, unless we assume that the average browsing bite rate is meaningful. Variation in observed rates (predicted ranges from 3 to 13 bites per min) suggests that species or genera variation is important.

```{r,echo=F, fig.align='center', fig.cap = 'Figure 9: Browser bite rates by genera and species. Points are median posterior predictions, with 95% certainty intervals.', fig.width= 8, fig.height = 5}

ggplot(pred.mean.genus, aes(class, median, col=class)) + geom_pointrange(aes(ymin = lower, ymax = upper)) + coord_flip() + theme(legend.position = 'NULL') + labs( x = '', y = 'Bites per minute', title = 'Browsers')

```
